Revised version of ERC-6123 including reference implementation for OTC-Derivatives and according Unit Tests
